Move BlangJson.cs to Classes folder
